requirements.txt
# data loading and manipulation
pandas

# data visualization
matplotlib
seaborn

# embeddings
sentence-transformers
transformers
langchain 


chromadb 
tqdm


# generative AI
google-generativeai

# interface
gradio